0.009
Dano implementado, animações de dano e morte são placeholds, dano meio bugado ainda, mas ta mais perto q longe

+++ click/Assets/Scripts/Towers/TowerData.cs
@@ -14,7 +14,7 @@ public class TowerData : ScriptableObject
     public GameObject gameModelTower;
     public GameObject gameModelTroop;
     public float atkSpeed;
-    public float atkDamage;
+    public int atkDamage;
     public string tipoRecurso;
     public int qtdRecurso;
